+++ HPXMLtoOpenStudio/tests/test_defaults.rb
@@ -381,9 +381,10 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.foundation_walls[0].interior_finish_type = HPXML::InteriorFinishGypsumCompositeBoard
     hpxml.foundation_walls[0].interior_finish_thickness = 0.625
     hpxml.foundation_walls[0].azimuth = 123
+    hpxml.foundation_walls[0].area = 789
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 7.0, HPXML::InteriorFinishGypsumCompositeBoard, 0.625, 123)
+    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 7.0, HPXML::InteriorFinishGypsumCompositeBoard, 0.625, 123, 789)
 
     # Test defaults
     hpxml.foundation_walls[0].thickness = nil
@@ -391,9 +392,11 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.foundation_walls[0].interior_finish_thickness = nil
     hpxml.foundation_walls[0].orientation = HPXML::OrientationSoutheast
     hpxml.foundation_walls[0].azimuth = nil
+    hpxml.foundation_walls[0].area = nil
+    hpxml.foundation_walls[0].length = 100
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 8.0, HPXML::InteriorFinishGypsumBoard, 0.5, 135)
+    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 8.0, HPXML::InteriorFinishGypsumBoard, 0.5, 135, 800)
 
     # Test defaults w/ unconditioned surfaces
     hpxml = _create_hpxml('base-foundation-unconditioned-basement.xml')
@@ -402,9 +405,12 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.foundation_walls[0].interior_finish_thickness = nil
     hpxml.foundation_walls[0].orientation = HPXML::OrientationSoutheast
     hpxml.foundation_walls[0].azimuth = nil
+    hpxml.foundation_walls[0].area = nil
+    hpxml.foundation_walls[0].length = 100
+    hpxml.foundation_walls[0].height = 10
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 8.0, HPXML::InteriorFinishNone, nil, 135)
+    _test_default_foundation_wall_values(hpxml_default.foundation_walls[0], 8.0, HPXML::InteriorFinishNone, nil, 135, 1000)
   end
 
   def test_frame_floors
@@ -2474,7 +2480,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     end
   end
 
-  def _test_default_foundation_wall_values(foundation_wall, thickness, int_finish_type, int_finish_thickness, azimuth)
+  def _test_default_foundation_wall_values(foundation_wall, thickness, int_finish_type, int_finish_thickness, azimuth, area)
     assert_equal(thickness, foundation_wall.thickness)
     assert_equal(int_finish_type, foundation_wall.interior_finish_type)
     if not int_finish_thickness.nil?
@@ -2483,6 +2489,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       assert_nil(foundation_wall.interior_finish_thickness)
     end
     assert_equal(azimuth, foundation_wall.azimuth)
+    assert_equal(area, foundation_wall.area)
   end
 
   def _test_default_frame_floor_values(frame_floor, int_finish_type, int_finish_thickness)